general: loads to connecting screen on python 3.x

+++ toontown/estate/GardenGlobals.py
@@ -298,7 +298,7 @@ if ACCELERATOR_USED_FROM_SHTIKER_BOOK:
     del PlantAttributes[202]
 
 def getTreeTrackAndLevel(typeIndex):
-    track = typeIndex / 7
+    track = int(typeIndex / 7)
     level = typeIndex % 7
     return (track, level)
